Cleanup of "main" programs, add a range for arrays
Make them ignore things used in unittest and verification test. Also
added an all target in Makefile. The array/pointer interface in Ranges.h
allows c++11 loops over old arrays.

+++ mach0/machin.cpp
@@ -1,5 +1,5 @@
 #include <cmath>
-double machin( double x, int i){
+double machin( double x, double i){
     double sign = pow(-1,i-1);
     double numerator = pow(x,2*i-1);
     double denominator = 2*i-1;
