+++ zeta0/zeta0.cpp
@@ -1,3 +1,3 @@
-double zetas2( int i ) {
+double zetas2( double i ) {
     return 1.0/i/i;
 }
